feat: create protected routes and verify response body. But missing fixed error at address field on POST

+++ src/modules/user/user.service.ts
@@ -1,7 +1,10 @@
-import { ConflictException, Injectable } from '@nestjs/common';
-import { CreateAddressDto, CreateUserDto } from './dto/create-user.dto';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { User } from './entities/user.entity';
 import { UserRepository } from './repositories/user.repository';
 
 @Injectable()
@@ -9,22 +12,26 @@ export class UserService {
   constructor(private userRepository: UserRepository) {}
 
   async create(createUserDto: CreateUserDto) {
-    const findUSer = await this.userRepository.findByEmail(createUserDto.email);
-    if (findUSer) {
-      throw new ConflictException('User already exists');
+    const findUser = await this.userRepository.findByEmail(createUserDto.email);
+    if (findUser) {
+      throw new ConflictException('User already exists!');
     }
     const user = await this.userRepository.create(createUserDto);
 
     return user;
   }
 
-  async findAll() {
-    const users = await this.userRepository.findAll();
+  async findAll(userLoggedId: string) {
+    const users = await this.userRepository.findAll(userLoggedId);
     return users;
   }
 
   async findOne(id: string) {
     const user = await this.userRepository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found!');
+    }
     return user;
   }
 
@@ -34,11 +41,21 @@ export class UserService {
   }
 
   async update(id: string, updateUserDto: UpdateUserDto) {
+    const user = await this.userRepository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found!');
+    }
     const newUser = await this.userRepository.update(id, updateUserDto);
     return newUser;
   }
 
   async remove(id: string) {
+    const user = await this.userRepository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found!');
+    }
     return await this.userRepository.remove(id);
   }
 }
